backend done

+++ jobs/views.py
@@ -1,14 +1,16 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, PermissionDenied
 from rest_framework import status
 
 from .models import Job
 from jobs.serializers.common import JobSerializer 
 from .serializers.populated import PopulatedJobsSerializer
 
+from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated
 
 class JobListView(APIView):
+  permission_classes = (IsAuthenticatedOrReadOnly, )
 
   def get(self, _request):
     jobs = Job.objects.all()
@@ -16,6 +18,8 @@ class JobListView(APIView):
     return Response(serialized_jobs.data, status.HTTP_200_OK)
 
   def post(self, request):
+      permission_classes = (IsAuthenticated, )
+      request.data['owner'] = request.user.id
       deserialized_job = JobSerializer(data=request.data)
       try:
         deserialized_job.is_valid(True)
@@ -27,6 +31,7 @@ class JobListView(APIView):
 
 
 class JobDetailListView(APIView):
+  permission_classes = (IsAuthenticated, )
 
   def get_job(self, pk):
     try:
@@ -39,16 +44,23 @@ class JobDetailListView(APIView):
     serialized_job = PopulatedJobsSerializer(job)
     return Response(serialized_job.data, status.HTTP_200_OK)
 
-  def delete(self, _request, pk):
+  def delete(self, request, pk):
     job_to_delete = self.get_job(pk)
+    if job_to_delete.owner != request.user:
+            print('WE CANT DELETE JOB')
+            raise PermissionDenied()
     job_to_delete.delete()
     return Response(status=status.HTTP_204_NO_CONTENT)
 
   def put(self, request, pk):
     job_to_update = self.get_job(pk=pk)
+    if job_to_update.owner != request.user:
+            print('WE CANT UPDATE JOB')
+            raise PermissionDenied()
+    # job_to_update.data['owner'] = request.user.id
     deserialized_job = JobSerializer(job_to_update, request.data)
     try:
-      deserialized_job.is_valid()
+      deserialized_job.is_valid(True)
       deserialized_job.save()
       return Response(deserialized_job.data, status.HTTP_202_ACCEPTED)
     except Exception as e:
